starting general cleaning and refactor - 2

+++ ex1/include/sorts.h
@@ -20,6 +20,6 @@ void insertionSort(void ** array, int size, int (*compare)(void*,void*));
 void mergeSort(void ** array, int leftIndex, int rightIndex, int (*compare)(void*,void*));
 void merge(void ** array, int left, int center, int right, int (*compare)(void*,void*));
 int sumsInArray(void ** array_to_test, void ** array_of_sums, int size_to_test, int size_of_sums,
-        int (*compare)(void*,void*), long long * (sums_support)(void*, void*));
+        int (*compare)(void*,void*), long * (sums_support)(void*, void*));
 
 #endif //SORTS_H_
